--- hradar/src/main/java/org/hit/hradar/approval/command/domain/aggregate/ApprovalActionType.java
@@ -1,9 +1,0 @@
-package org.hit.hradar.approval.command.domain.aggregate;
-
-public enum ApprovalActionType {
-
-  SUBMIT,   // 상신(제출)
-  APPROVE,  // 승인
-  REJECT,   // 거절
-  RECALL    // 회수
-}
