--- hradar/src/main/java/org/hit/hradar/approval/command/domain/aggregate/ApprovalStatus.java
@@ -1,10 +1,0 @@
-package org.hit.hradar.approval.command.domain.aggregate;
-
-public enum ApprovalStatus {
-
-  DRAFT,      // 임시저장
-  SUBMITTED,  // 상신(결재 진행 중)
-  APPROVED,   // 최종 승인
-  REJECTED,   // 반려
-  RECALL      // 회수
-}
