--- hradar/src/main/java/org/hit/hradar/attendance/command/domain/aggregate/AuthResult.java
@@ -1,8 +1,0 @@
-package org.hit.hradar.attendance.command.domain.aggregate;
-
-public enum AuthResult {
-
-  ALLOWED,    //허용 IP
-  DENIED      //허용하지 않는 IP
-
-}
