--- hradar/src/main/java/org/hit/hradar/attendance/command/domain/aggregate/Status.java
@@ -1,8 +1,0 @@
-package org.hit.hradar.attendance.command.domain.aggregate;
-
-public enum Status {
-
-  PENDING,    // 대기
-  APPROVED,   // 승인
-  REJECTED    // 반려
-}
