+++ hradar/src/main/java/org/hit/hradar/domain/approval/command/domain/aggregate/ApprovalLine.java
@@ -1,0 +1,27 @@
+package org.hit.hradar.domain.approval.command.domain.aggregate;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import org.hit.hradar.global.dto.BaseTimeEntity;
+
+@Entity
+@Table(name = "APPROVAL_LINE")
+@Getter
+public class ApprovalLine extends BaseTimeEntity {
+
+  //결재선id
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "approval_line_id")
+  private Long approvalLineId;
+
+  //결재문서id
+  @Column(name = "approval_document_id", nullable = false)
+  private Long approvalDocumentId;
+
+}
