+++ hradar/src/main/java/org/hit/hradar/domain/approval/command/domain/aggregate/ApprovalStepStatus.java
@@ -1,0 +1,9 @@
+package org.hit.hradar.domain.approval.command.domain.aggregate;
+
+public enum ApprovalStepStatus {
+
+  PENDING,    // 승인대기
+  APPROVED,   // 승인
+  REJECTED    // 반려
+
+}
