+++ hradar/src/main/java/org/hit/hradar/domain/attendance/command/application/service/AttendanceCommandService.java
@@ -1,0 +1,5 @@
+package org.hit.hradar.domain.attendance.command.application.service;
+
+public class AttendanceCommandService {
+
+}
