+++ hradar/src/main/java/org/hit/hradar/domain/attendance/command/domain/aggregate/AuthResult.java
@@ -1,0 +1,8 @@
+package org.hit.hradar.domain.attendance.command.domain.aggregate;
+
+public enum AuthResult {
+
+  ALLOWED,    //허용 IP
+  DENIED      //허용하지 않는 IP
+
+}
