+++ hradar/src/main/java/org/hit/hradar/domain/attendance/command/domain/aggregate/CorrectionType.java
@@ -1,0 +1,8 @@
+package org.hit.hradar.domain.attendance.command.domain.aggregate;
+
+public enum CorrectionType {
+
+  CHECK_IN_ERROR,   //출근 시간 오류
+  CHECK_OUT_ERROR   //퇴근 시간 오류
+
+}
