+++ hradar/src/main/java/org/hit/hradar/domain/attendance/command/domain/aggregate/Status.java
@@ -1,0 +1,8 @@
+package org.hit.hradar.domain.attendance.command.domain.aggregate;
+
+public enum Status {
+
+  PENDING,    // 대기
+  APPROVED,   // 승인
+  REJECTED    // 반려
+}
