+++ hradar/src/main/java/org/hit/hradar/domain/attendance/command/domain/aggregate/WorkType.java
@@ -1,0 +1,9 @@
+package org.hit.hradar.domain.attendance.command.domain.aggregate;
+
+public enum WorkType {
+
+  OFFICE,      // 내근
+  OUTSIDE,     // 외근
+  BUSINESS_TRIP, // 출장
+  REMOTE       // 재택
+}
