+++ hradar/src/main/java/org/hit/hradar/domain/holiday/command/domain/aggregate/HolidayType.java
@@ -1,0 +1,8 @@
+package org.hit.hradar.domain.holiday.command.domain.aggregate;
+
+public enum HolidayType {
+
+  LEGAL,        // 법정공휴일
+  COMPANY,      // 회사 지정 휴일
+  SUBSTITUTE    // 대체휴일
+}
