+++ hradar/src/main/java/org/hit/hradar/domain/leave/command/domain/aggregate/LeaveStatus.java
@@ -1,0 +1,9 @@
+package org.hit.hradar.domain.leave.command.domain.aggregate;
+
+public enum LeaveStatus {
+
+  REQUESTED,    // 신청
+  APPROVED,     // 승인
+  REJECTED,     // 반려
+  CANCELLED     // 취소
+}
