--- hradar/src/main/java/org/hit/hradar/leave/command/domain/aggregate/LeaveStatus.java
@@ -1,9 +1,0 @@
-package org.hit.hradar.leave.command.domain.aggregate;
-
-public enum LeaveStatus {
-
-  REQUESTED,    // 신청
-  APPROVED,     // 승인
-  REJECTED,     // 반려
-  CANCELLED     // 취소
-}
