--- hradar/src/main/java/org/hit/hradar/leave/command/domain/aggregate/LeaveType.java
@@ -1,12 +1,0 @@
-package org.hit.hradar.leave.command.domain.aggregate;
-
-public enum LeaveType {
-
-  ANNUAL,        // 연차
-  HALF_DAY,      // 반차 (0.5)
-  QUARTER_DAY,   // 반반차 (0.25)
-  FAMILY_EVENT,  // 경조사
-  SICK,          // 병가
-  OTHER          // 기타
-
-}
